Reorganize singularity engine configuration to get rid of non public struct fields and remove package dependencies which grow plugin binary size uselessly by almost 7 Mbytes.
